+++ Tools/HILSIM_XPlane/UDBSocket.h
@@ -29,7 +29,7 @@ typedef struct UDBSocket_t *UDBSocket;
 
 // Leave serial_port NULL and serial_baud 0 for UDP sockets
 // Leave UDP_port 0 for serial sockets
-UDBSocket UDBSocket_init(UDBSocketType type, long UDP_port, char *serial_port, long serial_baud);
+UDBSocket UDBSocket_init(UDBSocketType type, uint16_t UDP_port, char *serial_port, long serial_baud);
 
 void UDBSocket_close(UDBSocket socket);
